refactor: move tests near files
