+++ shared/utils/encrypt-values.test.ts
@@ -1,9 +1,5 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
-import {
-  encryptValue,
-  decryptValue,
-  decryptParts,
-} from '../../../../shared/utils/encrypt-values'
+import { encryptValue, decryptValue, decryptParts } from './encrypt-values'
 
 beforeEach(() => {
   vi.stubEnv('PR_HASH_SECRET', 'test-secret')
